+++ packages/pluto-py/CHANGELOG.md
@@ -1,5 +1,11 @@
 # pluto-client
 
+## 0.0.12
+
+feat(sdk): add ReactApp resource type
+
+Adds ReactApp resource type to support building and deploying React applications.
+
 ## 0.0.11
 
 feat(sdk): add `Secret` resource type